+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 CHANGELOG
 =========
 
+0.7
+---
+
+ * Add reranking support via `RerankingResult`, `RerankingEntry`, and `Capability::RERANKING`
+
 0.6
 ---
 
-* [BC BREAK] Change `Symfony\AI\Platform\Contract\JsonSchema\Factory` constructor signature in order to make schema generation extensible
+ * [BC BREAK] Change `Symfony\AI\Platform\Contract\JsonSchema\Factory` constructor signature in order to make schema generation extensible
 
 0.4
 ---